feat: Refactor test suite structure and add TestRunner for improved organization
